+++ src/io/flutter/performance/FlutterPerformanceViewFactory.java
@@ -1,9 +1,9 @@
 /*
- * Copyright 2018 The Chromium Authors. All rights reserved.
+ * Copyright 2019 The Chromium Authors. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
  */
-package io.flutter.view;
+package io.flutter.performance;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.components.ServiceManager;
@@ -12,9 +12,10 @@ import com.intellij.openapi.project.DumbService;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
+import io.flutter.view.FlutterViewMessages;
 import org.jetbrains.annotations.NotNull;
 
-public class FlutterPerfViewFactory implements ToolWindowFactory, DumbAware {
+public class FlutterPerformanceViewFactory implements ToolWindowFactory, DumbAware {
   @Override
   public void init(ToolWindow window) {
     window.setAvailable(true, null);
@@ -28,7 +29,7 @@ public class FlutterPerfViewFactory implements ToolWindowFactory, DumbAware {
 
   private static void initPerfView(@NotNull Project project, FlutterViewMessages.FlutterDebugEvent event) {
     ApplicationManager.getApplication().invokeLater(() -> {
-      final FlutterPerfView flutterPerfView = ServiceManager.getService(project, FlutterPerfView.class);
+      final FlutterPerformanceView flutterPerfView = ServiceManager.getService(project, FlutterPerformanceView.class);
       flutterPerfView.debugActive(event);
     });
   }
@@ -37,7 +38,7 @@ public class FlutterPerfViewFactory implements ToolWindowFactory, DumbAware {
   public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
     //noinspection CodeBlock2Expr
     DumbService.getInstance(project).runWhenSmart(() -> {
-      (ServiceManager.getService(project, FlutterPerfView.class)).initToolWindow(toolWindow);
+      (ServiceManager.getService(project, FlutterPerformanceView.class)).initToolWindow(toolWindow);
     });
   }
 }
