+++ src/io/flutter/view/DebugPaintAction.java
@@ -10,8 +10,8 @@ import io.flutter.run.daemon.FlutterApp;
 import io.flutter.vmService.ServiceExtensions;
 import org.jetbrains.annotations.NotNull;
 
-class DebugPaintAction extends FlutterViewToggleableAction {
-  DebugPaintAction(@NotNull FlutterApp app) {
+public class DebugPaintAction extends FlutterViewToggleableAction {
+  public DebugPaintAction(@NotNull FlutterApp app) {
     super(app, FlutterIcons.DebugPaint, ServiceExtensions.debugPaint);
   }
 }
